+++ server/src/tests/cases/compiler/buffer1.c
@@ -1,0 +1,5 @@
+// @driver: fluffos
+buffer b;
+test() {
+    buffer result = b + b;
+}