+++ server/src/tests/cases/compiler/bytes1.c
@@ -1,0 +1,5 @@
+bytes b;
+
+test() {
+    bytes result = b + b;
+}